ファイル一覧の管轄を修正。フォルダとかそっち系
