+++ src/features/core/extract/note/types.ts
@@ -3,13 +3,6 @@ export interface NoteSchema {
   kind: string;
 }
 
-export interface AudioFilesSet {
-  bgm: ReadonlySet<string>;
-  bgs: ReadonlySet<string>;
-  me: ReadonlySet<string>;
-  se: ReadonlySet<string>;
-}
-
 export interface XX {
   key: string;
   value: string;
@@ -25,16 +18,6 @@ export interface NoteSchema {
   key: string;
   kinds: string;
   source: string;
-}
-
-export interface ImageFilesSet {
-  characters: ReadonlySet<string>;
-  faces: ReadonlySet<string>;
-  svEnemy: ReadonlySet<string>;
-  svActors: ReadonlySet<string>;
-  enemies: ReadonlySet<string>;
-  picutures: ReadonlySet<string>;
-  tilesets: ReadonlySet<string>;
 }
 
 export interface OtherFilesSet {
