+++ src/fileio/audio/index.ts
@@ -1,1 +1,2 @@
 export * from "./constants";
+export * from "./types";
