+++ src/fileio/audio/types.ts
@@ -1,0 +1,6 @@
+export interface AudioFilesSet {
+  bgm: ReadonlySet<string>;
+  bgs: ReadonlySet<string>;
+  me: ReadonlySet<string>;
+  se: ReadonlySet<string>;
+}
