+++ src/fileio/image/types.ts
@@ -1,0 +1,9 @@
+export interface ImageFilesSet {
+  characters: ReadonlySet<string>;
+  faces: ReadonlySet<string>;
+  svEnemy: ReadonlySet<string>;
+  svActors: ReadonlySet<string>;
+  enemies: ReadonlySet<string>;
+  picutures: ReadonlySet<string>;
+  tilesets: ReadonlySet<string>;
+}
